+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Release Notes
+## [v1.0.6](https://github.com/404Exist/datatable-cruds/compare/v1.0.5...v1.0.6) - 2022-04-10
+
+### Added
+- `setBladeExtends()` and `setBladeSection()` methods have been added.
+### Updated
+- `render()` method now accepts one parameter only which is an array of variables that can be accessed in the layout blade file.
 ## [v1.0.5](https://github.com/404Exist/datatable-cruds/compare/v1.0.3...v1.0.5) - 2022-04-9
 
 ### Added